+++ src/mako/validator/rules/Optional.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\validator\rules;
+
+/**
+ * Optional rule.
+ *
+ * @author Frederic G. Østby
+ */
+class Optional extends Rule implements RuleInterface
+{
+	/**
+	 * {@inheritdoc}
+	 */
+	public function validate($value, array $input): bool
+	{
+		return true;
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getErrorMessage(string $field): string
+	{
+		return ''; // This rule can never fail so there isn't any error message
+	}
+}